make helper private and pull into own test file
